refactor: PC 버전 가게 생성 수정
