+++ src/page/ShopRegistration/component/Modal/ConfirmShop/index.tsx
@@ -1,36 +1,26 @@
-import useShopRegistrationStore from 'store/shopRegistration';
+import { useFormContext } from 'react-hook-form';
 import styles from './ConfirmShop.module.scss';
 
-interface ShopInfo {
-  name: string;
-  phone: string;
-  id: string;
-}
-
 interface ConfirmShopProps {
-  open: boolean;
-  selectedShop: ShopInfo;
   onCancel: () => void;
 }
 
-export default function ConfirmShop({ open, onCancel, selectedShop }: ConfirmShopProps) {
-  const { setName, setShopId } = useShopRegistrationStore();
-  if (!open) return null;
+export default function ConfirmShop({ onCancel }: ConfirmShopProps) {
+  const { getValues } = useFormContext();
+  const values = getValues();
   return (
     <div className={styles.container}>
       <div className={styles.container__info}>
-        <span className={styles.container__title}>{selectedShop.name}</span>
+        <span className={styles.container__title}>{values.name}</span>
         <div className={styles.container__text}>
           <span className={styles['container__phone-number-label']}>전화번호</span>
-          <span className={styles['container__phone-number']}>{selectedShop.phone}</span>
+          <span className={styles['container__phone-number']}>{values.phone}</span>
         </div>
       </div>
       <button
         type="button"
         className={styles['container__select-button']}
         onClick={() => {
-          setName(selectedShop.name);
-          setShopId(selectedShop.id);
           onCancel();
         }}
       >
