+++ src/page/ShopRegistration/view/PC/ShopConfirmation/ShopConfirmation.module.scss
@@ -1,64 +1,5 @@
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  appearance: none;
-  margin: 0;
-}
-
 .wrapper {
   position: relative;
-}
-
-.block {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  &__writing-icon {
-    margin-bottom: 72px;
-  }
-
-  &__title {
-    display: block;
-    text-align: center;
-    font-size: 36px;
-    font-weight: 700;
-    color: #175c8e;
-    margin-bottom: 24px;
-    height: 53px;
-  }
-
-  &__text {
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-    text-align: center;
-    font-size: 16px;
-    font-weight: 400;
-    margin-bottom: 80px;
-    color: #858585;
-
-    span {
-      line-height: 180%;
-    }
-  }
-
-  &__next-button {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 368px;
-    height: 48px;
-    font-weight: 500;
-    color: #ffffff;
-    background-color: #175c8e;
-    text-decoration: none;
-
-    &:hover {
-      cursor: pointer;
-    }
-  }
 }
 
 .container {
